CDAP-15847. Move packages and pom changes
